Refactor and simplify codebase for clarity and efficiency
- Removed `Google.Chrome.EXE` from `winget-config.json` and added `TortoiseGit.TortoiseGit`.
- Simplified type casting in `ProfitSharingUnder21BreakdownByStoreResponse.cs`, `ForfeituresAndPointsForYearEndpoint.cs`, and `FrozenReportServiceTests.cs`.
- Refactored `TerminatedEmployeeReportService.cs`:
  - Made `GetProfitYearRange` and `CombineEmployeeAndBeneficiarySlices` methods `static`.
  - Removed unused `GetTerminatedEmployees` and `GetBeneficiaries` methods.
  - Simplified `PSN` calculation logic.
- Replaced `DateTime.UtcNow` timing with `Stopwatch` in `TimingHelper.cs` for better precision.
- Removed unnecessary `(char?)` casting in `OracleEmployeeExtensions.cs`.
- Fixed BOM marker issues in `TerminatedEmployeeReportService.cs` and `AuditServiceTests.cs`.
- Simplified assertions in `AuditServiceTests.cs`.

These changes improve readability, maintainability, and performance.

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Response/PostFrozen/ProfitSharingUnder21BreakdownByStoreResponse.cs
@@ -26,7 +26,7 @@ public sealed record ProfitSharingUnder21BreakdownByStoreResponse : IIsExecutive
     {
         return new ProfitSharingUnder21BreakdownByStoreResponse()
         {
-            StoreNumber = (short)42,
+            StoreNumber = 42,
             BadgeNumber = 704234,
             FullName = "Johnson, Martha",
             BeginningBalance = 20014,
