+++ src/services/tests/Demoulas.ProfitSharing.UnitTests/Reports/YearEnd/FrozenReportServiceTests.cs
@@ -62,14 +62,14 @@ public class FrozenReportServiceTests : ApiTestBase<Program>
             pdArray[0].FederalTaxes = 0.5m;
             pdArray[0].StateTaxes = 0.25m;
 
-            pdArray[1].ProfitYear = (short)(testYear - 1);
+            pdArray[1].ProfitYear = testYear - 1;
             pdArray[1].ProfitCode = ProfitCode.Constants.IncomingContributions;
             pdArray[1].ProfitCodeId = ProfitCode.Constants.IncomingContributions.Id;
             pdArray[1].Contribution = 500m;
             pdArray[1].Earnings = 50m;
             pdArray[1].Forfeiture = 0m;
             pdArray[1].MonthToDate = 0;
-            pdArray[1].YearToDate = (short)(testYear - 1);
+            pdArray[1].YearToDate = testYear - 1;
             pdArray[1].FederalTaxes = 0.5m;
             pdArray[1].StateTaxes = 0.25m;
 
@@ -92,7 +92,7 @@ public class FrozenReportServiceTests : ApiTestBase<Program>
             ppArray[0].ProfitYear = testYear;
             ppArray[0].IncomeExecutive = 25;
             ppArray[0].CurrentIncomeYear = 49995;
-            ppArray[1].ProfitYear = (short)(testYear - 1);
+            ppArray[1].ProfitYear = testYear - 1;
             ppArray[1].IncomeExecutive = 0;
             ppArray[1].CurrentIncomeYear = 0;
